requirements.txt
streamlit>=1.43,<2
pandas>=2.2,<3
plotly>=6,<7
requests>=2.32,<3
PyYAML>=6.0,<7
duckdb>=1.2,<2
openpyxl>=3.1,<4
geopandas>=1.0,<2
fiona>=1.10,<2
shapely>=2.0,<3
pyarrow>=19,<20
pytest>=8.3,<9
